PROTOCOL-115 Move ConcurrentTestRunner to james-util
