+++ mailbox/pom.xml
@@ -215,6 +215,11 @@
                 <artifactId>apache-james-mailbox-tool</artifactId>
                 <version>${project.version}</version>
             </dependency>
+           <dependency>
+                <groupId>org.apache.james</groupId>
+                <artifactId>james-server-util</artifactId>
+                <version>${project.version}</version>
+            </dependency>
             <!--
                 END Modules
             -->
