+++ server/container/util/src/main/java/org/apache/james/util/concurrency/ConcurrentTestRunner.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.store.mail.model.concurrency;
+package org.apache.james.util.concurrency;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
